Implement ECDSA algorithm
